# Viterbi: numpy indexing is faster than unrolling numba