This PR speeds up pyin by about 23% (on a Mac M2, python 3.10).

My test call (run ten times):
```py
librosa.pyin(
            wav, # 528318 elements
            fmin=librosa.note_to_hz("C2"),
            fmax=librosa.note_to_hz("C5"),
            sr=16_000,
            frame_length=1024,
            hop_length=256,
            center=False
)
```

The argmax line is numba-compatible, but just having it with numba decreased speed by about 26%.