@bmcfee 2024-03-26
To summarize: this PR removes numba jitting from the core viterbi algorithm, and replaces an explicit loop with a vectorized call.

I'd be surprised if this is faster in general, though I can believe it will be faster in certain circumstances.  I'm curious how you're benchmarking it though?  Specifically, is your 23% over multiple trials, and not counting the initial jit compilation?

For a bit of context, I believe the vectorized argmax / fancy indexing was not numba-compatible at the time we implemented this (or with the minimum supported numba that we wanted to retain), hence the explicit for loop.