@twoertwein 2024-03-26
Good point, I did not exclude the warm up phase (I only averaged it over ten calls)! It might also be limited to Mac M2 (not sure how well numba works on that).

Here is a new benchmark (10 warmup rounds, followed by 100 rounds for timing):

```py
import time

import librosa
import numpy as np

wav = np.random.rand(528318)


# warm up
for i in range(10):
    librosa.pyin(
        wav,
        fmin=librosa.note_to_hz("C2"),
        fmax=librosa.note_to_hz("C5"),
        sr=16_000,
        frame_length=1024,
        hop_length=256,
        center=False,
    )


start = time.time()
for i in range(100):
    librosa.pyin(
        wav,
        fmin=librosa.note_to_hz("C2"),
        fmax=librosa.note_to_hz("C5"),
        sr=16_000,
        frame_length=1024,
        hop_length=256,
        center=False,
    )
print((time.time() - start) / 100)

# py 3.10, numpy: 1.26.4, numba: 0.59.1
# this PR on top of 0.10.1: 1.8619681692123413
# 0.10.1: 2.543446371555328
# ~26% faster
```