@bmcfee 2024-03-27
I just ran my own benchmark locally and performance is indistinguishable between the current main and this PR.

```python
import librosa
import numpy as np

y, sr = librosa.load(librosa.ex('trumpet'))

# Force a jit compilation if we didn't have it cached already
librosa.pyin(y=y, sr=sr, fmin=32, fmax=2000);

# Current main
%timeit librosa.pyin(y=y, sr=sr, fmin=32, fmax=2000)
# 1.43 s ± 22 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

# PR 1825
%timeit librosa.pyin(y=y, sr=sr, fmin=32, fmax=2000)
# 1.46 s ± 32 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

That's on the 5 second trumpet example.  On a longer example (nutcracker, 119 seconds), we get:

```
# Main
26.1 s ± 259 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

# PR 1825
27.3 s ± 266 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```