@twoertwein 2024-03-27
> I'd be surprised if this is faster in general, though I can believe it will be faster in certain circumstances. 

It would be great if you (and your collaborators) could run the following script on other platforms and other python version to see whether the pure numpy-version is worth it or not. Be-aware my results seem to be a bit noisy - sometimes having a larger range of F0 is faster than with a smaller range, which seems unexpected.

```py
import time

import librosa
import numpy as np

for length in (10_000, 100_000):
    for fmax in ("C8", "C7", "C6", "C5", "C4", "C3"):
        wav = np.random.rand(length)

        # warm up
        for i in range(10):
            librosa.pyin(
                wav, fmin=librosa.note_to_hz("C2"), fmax=librosa.note_to_hz(fmax)
            )

        start = time.time()
        for i in range(100):
            librosa.pyin(
                wav, fmin=librosa.note_to_hz("C2"), fmax=librosa.note_to_hz(fmax)
            )
        print(f"{length=}, {fmax=}", (time.time() - start) / 100)

# py 3.10, numpy: 1.26.4, numba: 0.59.1

# 0.10.1:
length=10000, fmax='C8' 0.15894932985305787
length=10000, fmax='C7' 0.1073629117012024
length=10000, fmax='C6' 0.0850246500968933
length=10000, fmax='C5' 0.05933178901672363
length=10000, fmax='C4' 0.02861675024032593
length=10000, fmax='C3' 0.012354247570037842
length=100000, fmax='C8' 1.0184268498420714
length=100000, fmax='C7' 0.6894018507003784
length=100000, fmax='C6' 0.5472757506370545
length=100000, fmax='C5' 0.3582835292816162
length=100000, fmax='C4' 0.20101286888122558
length=100000, fmax='C3' 0.0643980097770691

# this PR on top of 0.10.1:
length=10000, fmax='C8' 0.11170340061187745
length=10000, fmax='C7' 0.0782641363143921
length=10000, fmax='C6' 0.07668728113174439
length=10000, fmax='C5' 0.04809794187545777
length=10000, fmax='C4' 0.0173807692527771
length=10000, fmax='C3' 0.011447958946228028
length=100000, fmax='C8' 0.598180022239685
length=100000, fmax='C7' 0.41540180921554565
length=100000, fmax='C6' 0.4820923900604248
length=100000, fmax='C5' 0.3346623587608337
length=100000, fmax='C4' 0.1822682499885559
length=100000, fmax='C3' 0.05109306097030639
```

> I believe the vectorized argmax / fancy indexing was not numba-compatible at the time we implemented this

The argmax seems to be compatible now but the fancy indexing is still not supported.

Using just the argmax+jit (and the indexing in the loop) - surprisingly it becomes slower:
```py
length=10000, fmax='C8' 0.197688729763031
length=10000, fmax='C7' 0.14093083143234253
length=10000, fmax='C6' 0.10329761981964111
length=10000, fmax='C5' 0.062146799564361574
length=10000, fmax='C4' 0.02347614288330078
length=10000, fmax='C3' 0.012690157890319824
length=100000, fmax='C8' 1.372423028945923
length=100000, fmax='C7' 0.9433927917480469
length=100000, fmax='C6' 0.6792442488670349
length=100000, fmax='C5' 0.4484334897994995
length=100000, fmax='C4' 0.26358938932418824
length=100000, fmax='C3' 0.06613000869750976
```